requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
pycoingecko
python-multipart
